fix(sequence): add steps in sequence.ts instead of individual files. Closes #2
