+++ src/application.ts
@@ -10,8 +10,6 @@ import {ServiceMixin} from '@loopback/service-proxy';
 import dotenv from 'dotenv';
 import dotEnvExtended from 'dotenv-extended';
 import path from 'path';
-import {OriginCheckSequence} from './origin-check.sequence';
-import {RequestLoggerSequence} from './request-logger.sequence';
 import {MySequence} from './sequence';
 
 export {ApplicationConfig};
@@ -28,8 +26,6 @@ export class UserCrudApplication extends BootMixin(
 
     // Set up the custom sequence
     this.sequence(MySequence);
-    this.sequence(RequestLoggerSequence);
-    this.sequence(OriginCheckSequence);
 
     // Set up default home page
     this.static('/', path.join(__dirname, '../public'));
